+++ datamations/__init__.py
@@ -3,11 +3,10 @@
 # Create a plot datamation from a pandas pipeline.
 #
 
-from .datamation_sanddance import datamation_sanddance
 from .small_salary import small_salary
-from .DatamationFrame import DatamationFrame
-from .DatamationGroupBy import DatamationGroupBy
+from .datamation_frame import DatamationFrame
+from .datamation_groupby import DatamationGroupBy
 
 __all__ = [
-    "small_salary", "datamation_sanddance", "DatamationFrame", "DatamationGroupBy"
+    "small_salary", "DatamationFrame", "DatamationGroupBy"
 ]
